+++ src/game/Object/Player.cpp
@@ -3938,327 +3938,6 @@ void Player::SetHover(bool enable)
     SendMessageToSet(&data, true);
 }
 
-/* Preconditions:
-  - a resurrectable corpse must not be loaded for the player (only bones)
-  - the player must be in world
-*/
-void Player::BuildPlayerRepop()
-{
-    if (getRace() == RACE_NIGHTELF)
-    {
-        CastSpell(this, 20584, true); // auras SPELL_AURA_INCREASE_SPEED(+speed in wisp form), SPELL_AURA_INCREASE_SWIM_SPEED(+swim speed in wisp form), SPELL_AURA_TRANSFORM (to wisp form)
-    }
-    CastSpell(this, 8326, true);                            // auras SPELL_AURA_GHOST, SPELL_AURA_INCREASE_SPEED(why?), SPELL_AURA_INCREASE_SWIM_SPEED(why?)
-
-    // the player can not have a corpse already, only bones which are not returned by GetCorpse
-    if (GetCorpse())
-    {
-        sLog.outError("BuildPlayerRepop: player %s(%d) already has a corpse", GetName(), GetGUIDLow());
-        MANGOS_ASSERT(false);
-    }
-
-    // create a corpse and place it at the player's location
-    Corpse* corpse = CreateCorpse();
-    if (!corpse)
-    {
-        sLog.outError("Error creating corpse for Player %s [%u]", GetName(), GetGUIDLow());
-        return;
-    }
-    GetMap()->Add(corpse);
-
-    // convert player body to ghost
-    SetHealth(1);
-
-    SetWaterWalk(true);
-    if (!GetSession()->isLogingOut())
-    {
-        SetRoot(false);
-    }
-
-    // BG - remove insignia related
-    RemoveFlag(UNIT_FIELD_FLAGS, UNIT_FLAG_SKINNABLE);
-
-    SendCorpseReclaimDelay();
-
-    // to prevent cheating
-    corpse->ResetGhostTime();
-
-    StopMirrorTimers();                                     // disable timers(bars)
-
-    // set and clear other
-    SetByteValue(UNIT_FIELD_BYTES_1, 3, UNIT_BYTE1_FLAG_ALWAYS_STAND);
-}
-
-/**
- * @brief Restores the player to life and reapplies post-resurrection effects.
- *
- * @param restore_percent The fraction of health and resources to restore.
- * @param applySickness True to apply resurrection sickness when appropriate.
- */
-void Player::ResurrectPlayer(float restore_percent, bool applySickness)
-{
-    WorldPacket data(SMSG_DEATH_RELEASE_LOC, 4 * 4);        // remove spirit healer position
-    data << uint32(-1);
-    data << float(0);
-    data << float(0);
-    data << float(0);
-    GetSession()->SendPacket(&data);
-
-    // speed change, land walk
-
-    // remove death flag + set aura
-    SetByteValue(UNIT_FIELD_BYTES_1, 3, 0x00);
-
-    SetDeathState(ALIVE);
-
-    if (getRace() == RACE_NIGHTELF)
-    {
-        RemoveAurasDueToSpell(20584); // speed bonuses
-    }
-    RemoveAurasDueToSpell(8326);                            // SPELL_AURA_GHOST
-
-    SetWaterWalk(false);
-    SetRoot(false);
-
-    // set health/powers (0- will be set in caller)
-    if (restore_percent > 0.0f)
-    {
-        SetHealth(uint32(GetMaxHealth()*restore_percent));
-        SetPower(POWER_MANA, uint32(GetMaxPower(POWER_MANA)*restore_percent));
-        SetPower(POWER_RAGE, 0);
-        SetPower(POWER_ENERGY, uint32(GetMaxPower(POWER_ENERGY)*restore_percent));
-    }
-
-    // trigger update zone for alive state zone updates
-    uint32 newzone, newarea;
-    GetZoneAndAreaId(newzone, newarea);
-    UpdateZone(newzone, newarea);
-
-    m_deathTimer = 0;
-
-    // update visibility of world around viewpoint
-    m_camera.UpdateVisibilityForOwner();
-    // update visibility of player for nearby cameras
-    UpdateObjectVisibility();
-
-#ifdef ENABLE_ELUNA
-    if (Eluna* e = GetEluna())
-    {
-        e->OnResurrect(this);
-    }
-#endif /* ENABLE_ELUNA */
-
-    if (!applySickness)
-    {
-        return;
-    }
-
-    // Characters from level 1-10 are not affected by resurrection sickness.
-    // Characters from level 11-19 will suffer from one minute of sickness
-    // for each level they are above 10.
-    // Characters level 20 and up suffer from ten minutes of sickness.
-    int32 startLevel = sWorld.getConfig(CONFIG_INT32_DEATH_SICKNESS_LEVEL);
-
-    if (int32(getLevel()) >= startLevel)
-    {
-        // set resurrection sickness
-        CastSpell(this, SPELL_ID_PASSIVE_RESURRECTION_SICKNESS, true);
-
-        // not full duration
-        if (int32(getLevel()) < startLevel + 9)
-        {
-            int32 delta = (int32(getLevel()) - startLevel + 1) * MINUTE;
-
-            if (SpellAuraHolder* holder = GetSpellAuraHolder(SPELL_ID_PASSIVE_RESURRECTION_SICKNESS))
-            {
-                holder->SetAuraDuration(delta * IN_MILLISECONDS);
-                holder->UpdateAuraDuration();
-            }
-        }
-    }
-}
-
-/**
- * @brief Transitions the player into the corpse state after death.
- */
-void Player::KillPlayer()
-{
-    SetRoot(true);
-
-    StopMirrorTimers();                                     // disable timers(bars)
-
-    SetDeathState(CORPSE);
-    // SetFlag( UNIT_FIELD_FLAGS, UNIT_FLAG_NOT_IN_PVP );
-
-    SetUInt32Value(UNIT_DYNAMIC_FLAGS, UNIT_DYNFLAG_NONE);
-    ApplyModByteFlag(PLAYER_FIELD_BYTES, 0, PLAYER_FIELD_BYTE_RELEASE_TIMER, !sMapStore.LookupEntry(GetMapId())->Instanceable());
-
-    // 6 minutes until repop at graveyard
-    m_deathTimer = 6 * MINUTE * IN_MILLISECONDS;
-
-    UpdateCorpseReclaimDelay();                             // dependent at use SetDeathPvP() call before kill
-
-    // don't create corpse at this moment, player might be falling
-
-    // update visibility
-    UpdateObjectVisibility();
-}
-
-/**
- * @brief Creates a corpse object for the player's current death state.
- *
- * @return The created corpse, or null if creation failed.
- */
-Corpse* Player::CreateCorpse()
-{
-    // prevent existence 2 corpse for player
-    SpawnCorpseBones();
-
-    Corpse* corpse = new Corpse((m_ExtraFlags & PLAYER_EXTRA_PVP_DEATH) ? CORPSE_RESURRECTABLE_PVP : CORPSE_RESURRECTABLE_PVE);
-    SetPvPDeath(false);
-
-    if (!corpse->Create(sObjectMgr.GenerateCorpseLowGuid(), this))
-    {
-        delete corpse;
-        return NULL;
-    }
-
-    uint8 skin       = GetByteValue(PLAYER_BYTES, 0);
-    uint8 face       = GetByteValue(PLAYER_BYTES, 1);
-    uint8 hairstyle  = GetByteValue(PLAYER_BYTES, 2);
-    uint8 haircolor  = GetByteValue(PLAYER_BYTES, 3);
-    uint8 facialhair = GetByteValue(PLAYER_BYTES_2, 0);
-
-    corpse->SetByteValue(CORPSE_FIELD_BYTES_1, 1, getRace());
-    corpse->SetByteValue(CORPSE_FIELD_BYTES_1, 2, getGender());
-    corpse->SetByteValue(CORPSE_FIELD_BYTES_1, 3, skin);
-
-    corpse->SetByteValue(CORPSE_FIELD_BYTES_2, 0, face);
-    corpse->SetByteValue(CORPSE_FIELD_BYTES_2, 1, hairstyle);
-    corpse->SetByteValue(CORPSE_FIELD_BYTES_2, 2, haircolor);
-    corpse->SetByteValue(CORPSE_FIELD_BYTES_2, 3, facialhair);
-
-    uint32 flags = CORPSE_FLAG_UNK2;
-    if (HasFlag(PLAYER_FLAGS, PLAYER_FLAGS_HIDE_HELM))
-    {
-        flags |= CORPSE_FLAG_HIDE_HELM;
-    }
-    if (HasFlag(PLAYER_FLAGS, PLAYER_FLAGS_HIDE_CLOAK))
-    {
-        flags |= CORPSE_FLAG_HIDE_CLOAK;
-    }
-    if (InBattleGround() && !InArena())
-    {
-        flags |= CORPSE_FLAG_LOOTABLE; // to be able to remove insignia
-    }
-    corpse->SetUInt32Value(CORPSE_FIELD_FLAGS, flags);
-
-    corpse->SetUInt32Value(CORPSE_FIELD_DISPLAY_ID, GetNativeDisplayId());
-
-    corpse->SetUInt32Value(CORPSE_FIELD_GUILD, GetGuildId());
-
-    uint32 iDisplayID;
-    uint32 iIventoryType;
-    uint32 _cfi;
-    for (int i = 0; i < EQUIPMENT_SLOT_END; ++i)
-    {
-        if (m_items[i])
-        {
-            iDisplayID = m_items[i]->GetProto()->DisplayInfoID;
-            iIventoryType = m_items[i]->GetProto()->InventoryType;
-
-            _cfi =  iDisplayID | (iIventoryType << 24);
-            corpse->SetUInt32Value(CORPSE_FIELD_ITEM + i, _cfi);
-        }
-    }
-
-    // we not need saved corpses for BG/arenas
-    if (!GetMap()->IsBattleGroundOrArena())
-    {
-        corpse->SaveToDB();
-    }
-
-    // register for player, but not show
-    sObjectAccessor.AddCorpse(corpse);
-    return corpse;
-}
-
-/**
- * @brief Converts an existing corpse into bones and persists the ghost state if needed.
- */
-void Player::SpawnCorpseBones()
-{
-    if (sObjectAccessor.ConvertCorpseForPlayer(GetObjectGuid()))
-        if (!GetSession()->PlayerLogoutWithSave())          // at logout we will already store the player
-        {
-            SaveToDB(); // prevent loading as ghost without corpse
-        }
-}
-
-/**
- * @brief Gets the player's current corpse object.
- *
- * @return The player's corpse, or null if none exists.
- */
-Corpse* Player::GetCorpse() const
-{
-    return sObjectAccessor.GetCorpseForPlayerGUID(GetObjectGuid());
-}
-
-/**
- * @brief Moves the player to the nearest valid graveyard.
- */
-void Player::RepopAtGraveyard()
-{
-    // note: this can be called also when the player is alive
-    // for example from WorldSession::HandleMovementOpcodes
-
-    AreaTableEntry const* zone = GetAreaEntryByAreaID(GetAreaId());
-
-    // Such zones are considered unreachable as a ghost and the player must be automatically revived
-    if ((!IsAlive() && zone && zone->flags & AREA_FLAG_NEED_FLY) || GetTransport())
-    {
-        ResurrectPlayer(0.5f);
-        SpawnCorpseBones();
-    }
-
-    WorldSafeLocsEntry const* ClosestGrave = NULL;
-
-    // Special handle for battleground maps
-    if (BattleGround* bg = GetBattleGround())
-    {
-        ClosestGrave = bg->GetClosestGraveYard(this);
-    }
-    else
-    {
-        ClosestGrave = sObjectMgr.GetClosestGraveYard(GetPositionX(), GetPositionY(), GetPositionZ(), GetMapId(), GetTeam());
-    }
-
-    // stop countdown until repop
-    m_deathTimer = 0;
-
-    // if no grave found, stay at the current location
-    // and don't show spirit healer location
-    if (ClosestGrave)
-    {
-        bool updateVisibility = IsInWorld() && GetMapId() == ClosestGrave->map_id;
-        TeleportTo(ClosestGrave->map_id, ClosestGrave->x, ClosestGrave->y, ClosestGrave->z, GetOrientation());
-        if (IsDead())                                       // not send if alive, because it used in TeleportTo()
-        {
-            WorldPacket data(SMSG_DEATH_RELEASE_LOC, 4 * 4);// show spirit healer position on minimap
-            data << ClosestGrave->map_id;
-            data << ClosestGrave->x;
-            data << ClosestGrave->y;
-            data << ClosestGrave->z;
-            GetSession()->SendPacket(&data);
-        }
-        if (updateVisibility && IsInWorld())
-        {
-            UpdateVisibilityAndView();
-        }
-    }
-}
-
 /**
  * @brief Registers a joined chat channel on the player.
  *
@@ -8116,220 +7795,6 @@ void Player::ApplyEquipCooldown(Item* pItem)
         data << ObjectGuid(pItem->GetObjectGuid());
         data << uint32(spellData.SpellId);
         GetSession()->SendPacket(&data);
-    }
-}
-
-/**
- * @brief Resets the player's learned spells and restores default and quest rewards.
- */
-void Player::resetSpells()
-{
-    // not need after this call
-    if (HasAtLoginFlag(AT_LOGIN_RESET_SPELLS))
-    {
-        RemoveAtLoginFlag(AT_LOGIN_RESET_SPELLS, true);
-    }
-
-    // make full copy of map (spells removed and marked as deleted at another spell remove
-    // and we can't use original map for safe iterative with visit each spell at loop end
-    PlayerSpellMap smap = GetSpellMap();
-
-    for (PlayerSpellMap::const_iterator iter = smap.begin(); iter != smap.end(); ++iter)
-    {
-        removeSpell(iter->first, false, false); // only iter->first can be accessed, object by iter->second can be deleted already
-    }
-
-    learnDefaultSpells();
-    learnQuestRewardedSpells();
-}
-
-/**
- * @brief Teaches the player's default race and class spells.
- */
-void Player::learnDefaultSpells()
-{
-    // learn default race/class spells
-    PlayerInfo const* info = sObjectMgr.GetPlayerInfo(getRace(), getClass());
-    for (PlayerCreateInfoSpells::const_iterator itr = info->spell.begin(); itr != info->spell.end(); ++itr)
-    {
-        uint32 tspell = *itr;
-        DEBUG_LOG("PLAYER (Class: %u Race: %u): Adding initial spell, id = %u", uint32(getClass()), uint32(getRace()), tspell);
-        if (!IsInWorld())                                   // will send in INITIAL_SPELLS in list anyway at map add
-        {
-            addSpell(tspell, true, true, true, false);
-        }
-        else                                                // but send in normal spell in game learn case
-        {
-            learnSpell(tspell, true);
-        }
-    }
-}
-
-/**
- * @brief Teaches a quest reward spell if the quest grants a learnable spell.
- *
- * @param quest The rewarded quest to inspect.
- */
-void Player::learnQuestRewardedSpells(Quest const* quest)
-{
-    uint32 spell_id = quest->GetRewSpellCast();
-
-    // skip quests without rewarded spell
-    if (!spell_id)
-    {
-        return;
-    }
-
-    SpellEntry const* spellInfo = sSpellStore.LookupEntry(spell_id);
-    if (!spellInfo)
-    {
-        return;
-    }
-
-    // check learned spells state
-    bool found = false;
-    for (int i = 0; i < MAX_EFFECT_INDEX; ++i)
-    {
-        if (spellInfo->Effect[i] == SPELL_EFFECT_LEARN_SPELL && !HasSpell(spellInfo->EffectTriggerSpell[i]))
-        {
-            found = true;
-            break;
-        }
-    }
-
-    // skip quests with not teaching spell or already known spell
-    if (!found)
-    {
-        return;
-    }
-
-    // prevent learn non first rank unknown profession and second specialization for same profession)
-    uint32 learned_0 = spellInfo->EffectTriggerSpell[EFFECT_INDEX_0];
-    if (sSpellMgr.GetSpellRank(learned_0) > 1 && !HasSpell(learned_0))
-    {
-        // not have first rank learned (unlearned prof?)
-        uint32 first_spell = sSpellMgr.GetFirstSpellInChain(learned_0);
-        if (!HasSpell(first_spell))
-        {
-            return;
-        }
-
-        SpellEntry const* learnedInfo = sSpellStore.LookupEntry(learned_0);
-        if (!learnedInfo)
-        {
-            return;
-        }
-
-        // specialization
-        if (learnedInfo->Effect[EFFECT_INDEX_0] == SPELL_EFFECT_TRADE_SKILL && learnedInfo->Effect[EFFECT_INDEX_1] == 0)
-        {
-            // search other specialization for same prof
-            for (PlayerSpellMap::const_iterator itr = m_spells.begin(); itr != m_spells.end(); ++itr)
-            {
-                if (itr->second.state == PLAYERSPELL_REMOVED || itr->first == learned_0)
-                {
-                    continue;
-                }
-
-                SpellEntry const* itrInfo = sSpellStore.LookupEntry(itr->first);
-                if (!itrInfo)
-                {
-                    return;
-                }
-
-                // compare only specializations
-                if (itrInfo->Effect[EFFECT_INDEX_0] != SPELL_EFFECT_TRADE_SKILL || itrInfo->Effect[EFFECT_INDEX_1] != 0)
-                {
-                    continue;
-                }
-
-                // compare same chain spells
-                if (sSpellMgr.GetFirstSpellInChain(itr->first) != first_spell)
-                {
-                    continue;
-                }
-
-                // now we have 2 specialization, learn possible only if found is lesser specialization rank
-                if (!sSpellMgr.IsHighRankOfSpell(learned_0, itr->first))
-                {
-                    return;
-                }
-            }
-        }
-    }
-
-    CastSpell(this, spell_id, true);
-}
-
-/**
- * @brief Reapplies all quest reward spells from previously rewarded quests.
- */
-void Player::learnQuestRewardedSpells()
-{
-    // learn spells received from quest completing
-    for (QuestStatusMap::const_iterator itr = mQuestStatus.begin(); itr != mQuestStatus.end(); ++itr)
-    {
-        // skip no rewarded quests
-        if (!itr->second.m_rewarded)
-        {
-            continue;
-        }
-
-        Quest const* quest = sObjectMgr.GetQuestTemplate(itr->first);
-        if (!quest)
-        {
-            continue;
-        }
-
-        learnQuestRewardedSpells(quest);
-    }
-}
-
-/**
- * @brief Learns or removes spells unlocked by a profession or skill value.
- *
- * @param skill_id The skill line identifier.
- * @param skill_value The current skill value.
- */
-void Player::learnSkillRewardedSpells(uint32 skill_id, uint32 skill_value)
-{
-    uint32 raceMask  = getRaceMask();
-    uint32 classMask = getClassMask();
-    for (uint32 j = 0; j < sSkillLineAbilityStore.GetNumRows(); ++j)
-    {
-        SkillLineAbilityEntry const* pAbility = sSkillLineAbilityStore.LookupEntry(j);
-        if (!pAbility || pAbility->skillId != skill_id || pAbility->learnOnGetSkill != ABILITY_LEARNED_ON_GET_PROFESSION_SKILL)
-        {
-            continue;
-        }
-        // Check race if set
-        if (pAbility->racemask && !(pAbility->racemask & raceMask))
-        {
-            continue;
-        }
-        // Check class if set
-        if (pAbility->classmask && !(pAbility->classmask & classMask))
-        {
-            continue;
-        }
-
-        if (sSpellStore.LookupEntry(pAbility->spellId))
-        {
-            // need unlearn spell
-            if (skill_value < pAbility->req_skill_value)
-            {
-                removeSpell(pAbility->spellId);
-            }
-            // need learn
-            else if (!IsInWorld())
-            {
-                addSpell(pAbility->spellId, true, true, true, false);
-            }
-            else
-            {
-                learnSpell(pAbility->spellId, true);
-            }
-        }
     }
 }
